[split] Extract code into separate rpc client (#1433)
Refactor command client so it can be used by multiple applications at
once. I need this for my clipboard manager that is active at the same
time as vscode.

This pull request basically does two things:
1. Splits the very large `command_client.py` in to several python files
in the new `rpc_client` directory. This is done purely mechanically with
as minimal changes as possible.
2. Introduces a new Talon action `rpc_client_run_command` that takes the
communication directory name as well as the callback to press the keys
that starts the request execution. This is what allows us to run
multiple applications at once.

---------

Co-authored-by: Nicholas Riley <com-github@sabi.net>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>